+++ frontend/src/routes/models/[id]/+page.svelte
@@ -800,6 +800,7 @@
 								<EntityDetailPanel
 									entity={selectedBrowseNode.data}
 									onclose={() => (selectedBrowseNode = null)}
+									currentModelId={model?.id}
 								/>
 							</div>
 						{/if}